Listen to keys, scalable to any height.
This commit deletes all previous versions of the files. Given that I'm
using git, I might as well use it. The process so far has been to open a
new project, copy the previous project in its entirety and edit to make
new changes where the file name represents the summary of the addition.
This is ridiculous. I might as well do it the right way and use the
commit title as the summary and keep it all in one file rather than
having a mine field of previous versions.