Add wind effect on a frame timer
Wind is randomly generated at random intervals of fromes passing,
between 1 and 6 seconds. This should be enough to not throw the same
wind for the same fall twice. In addition, since the wind adds itself to
the velocity of the character ball, it can switch directions mid fall
without throwing the ball in an unrealistic fashion