Align target accuracy and launch height
This doesn't make the launch height based on how accurate the ball is to
the bull's eye. Rather it dampens the wind based on the launch, which
make the ball go closer to the bull's eye. This is sort of an illusion
which will intentionally confuse the user to think that the wind must
play into the equation (although it is merely a special effect).